requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Homomorphic Encryption
tenseal>=0.3.14

# Data processing
scipy>=1.7.0

# Visualization & Analysis
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
